requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
Django==4.1.7
django-crispy-forms==2.0
django-markdown2==0.3.1
djangorestframework==3.14.0
markdown2==2.4.8
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
utils==1.0.1
django-on-heroku==1.1.2
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0